Python Set
A Python set is an unordered collection of unique elements. It is mutable, allowing you to add or remove items. Sets are useful for eliminating duplicates and performing operations like union, intersection, and difference.
Key Points:
Unordered: No specific order of elements.
Unique: Duplicate elements are not allowed.
Mutable: Elements can be added or removed.